api/requirements.txt
pinecone
pinecone_text
langchain_openai
langchain
dotenv
requests
pinecone[grpc]
fastapi[standard]
uvicorn[standard]
locust
google-cloud-firestore